feat(job): refactor kube_job
